refactor(vector): CompileFilter free function; move Filter into vtypes
